refactor: change all money-related fields to big decimal data type

+++ src/main/java/com/dvk/ct250backend/domain/booking/dto/BookingPriceDetailDTO.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -16,5 +18,5 @@ public class BookingPriceDetailDTO {
 
   String passengerType;
 
-  Double priceAmount;
+  BigDecimal priceAmount;
 }
